mover rutas y viewsets a proyecto
