+++ back-end/django/enbalde/urls.py
@@ -1,9 +1,34 @@
-from django.urls import path
+from django.urls import path, include
+from rest_framework import routers, viewsets
+from .models import Usuario, Articulo, TipoArticulo
+from .serializers import UsuarioSerializer, ArticuloSerializer, TipoArticuloSerializer
 from .views.usuario_views import LoginView, LogoutView, SignupView
 from .views.articulo_views import MuchosArticulos, UnArticulo
 from .views.tipo_articulo_views import MuchosTiposArticulos, UnTipoArticulo
+from .models import Usuario
+
+
+class UsuarioViewSet(viewsets.ModelViewSet):
+    queryset = Usuario.objects.all()
+    serializer_class = UsuarioSerializer
+
+
+class ArticuloViewSet(viewsets.ModelViewSet):
+    queryset = Articulo.objects.all()
+    serializer_class = ArticuloSerializer
+
+
+class TipoArticuloViewSet(viewsets.ModelViewSet):
+    queryset = TipoArticulo.objects.all()
+    serializer_class = TipoArticuloSerializer
+
+
+router = routers.DefaultRouter()
+router.register('usuarios', UsuarioViewSet)
+
 
 urlpatterns = [
+    path('', include(router.urls)),
     path('auth/login/', LoginView.as_view(), name='auth_login'),
     path('auth/logout/', LogoutView.as_view(), name='auth_logout'),
     path('auth/signup/', SignupView.as_view(), name='auth_signup'),
